refactor(ntfctl): streamline notification fetching and utils
- Remove manual osascript handling in latest-notification and notification-count
- Introduce fetchNotifications utility for cleaner code
- Update package dependencies for compatibility and performance
- Enhance .gitignore for project-specific files
- Update ESLint configuration file
