+++ extensions/ntfctl/.gitignore
@@ -1,4 +1,15 @@
-node_modules/
-dist/
-*.scpt
+# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
+
+# dependencies
+/node_modules
+
+# Raycast specific files
+dist
+raycast-env.d.ts
+.raycast-swift-build
+.swiftpm
+compiled_raycast_swift
+compiled_raycast_rust
+
+# misc
 .DS_Store
