+++ extensions/ntfctl/package.json
@@ -68,15 +68,15 @@
     }
   ],
   "dependencies": {
-    "@raycast/api": "^1.94.0"
+    "@raycast/api": "^1.104.23"
   },
   "devDependencies": {
-    "@raycast/eslint-config": "^2.0.4",
-    "@types/node": "22.13.10",
-    "@types/react": "^19.0.0",
-    "eslint": "^8.57.1",
-    "prettier": "^3.5.3",
-    "typescript": "^5.8.2"
+    "@raycast/eslint-config": "^2.2.0",
+    "@types/node": "26.1.1",
+    "@types/react": "^19.2.17",
+    "eslint": "^10.7.0",
+    "prettier": "^3.9.5",
+    "typescript": "^6.0.0"
   },
   "scripts": {
     "build": "ray build -e dist -o dist",
